requirements.txt
flask
Flask-Cors
scikit-learn
nltk
pandas
numpy
matplotlib
seaborn
joblib
python-dotenv
requests
beautifulsoup4
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
google-auth
Werkzeug
